In upload view: added zip and redirection.
	* Users can now upload zip archives through the upload_submission view.
	* Users will be redirected back to their previous page if they
	  traveled to the upload_submission view through their web browser, or
	  a json object will be sent if the view was requested through AJAX.

+++ components/web/galahweb/views/_upload_submission.py
@@ -4,8 +4,10 @@ from flask.ext.login import current_user
 from galahweb.auth import account_type_required
 from bson.objectid import ObjectId
 from bson.errors import InvalidId
-from flask import abort, render_template, request
+from flask import abort, render_template, request, flash, redirect, jsonify, \
+                  url_for
 from galah.db.models import Submission, Assignment
+from galahweb.util import is_url_on_site
 import os.path
 import subprocess
 import datetime
@@ -13,6 +15,47 @@ import shutil
 import tempfile
 
 SUBMISSION_DIRECTORY = "/var/local/galah.web/submissions/"
+assert SUBMISSION_DIRECTORY[0] == "/" # Directory must be given as absolute path
+
+def prepare_new_submission(**kwargs):
+    """
+    Prepares a new submission object properly initialized with an id and a
+    new testables directory to store the submission in.
+
+    """
+    
+    # Create a new submission from the keyword arguments
+    new_submission = Submission(**kwargs)
+    
+    # Create an id for the new submission if one doesn't yet exist
+    new_submission.id = new_submission.id or ObjectId()
+    
+    if not new_submission.testables:
+        # Craft a unique directory path where we will store the new submission
+        new_submission.testables = os.path.join(
+            SUBMISSION_DIRECTORY, str(new_submission.id)
+        )
+        
+        # Create the directory. We are guarenteed an ObjectId is unique. However
+        # we are not guarenteed that we will have the proper permissions and
+        # that we will be able to make the directory thus this could error
+        # because of that.
+        os.makedirs(new_submission.testables)
+        
+    return new_submission
+    
+def abort_new_submission(submission):
+    """
+    Deletes and cleans up a submission THAT HAS NOT YET BEEN SAVED to Mongo by
+    deleting the testables directory, etc.
+    
+    Call this if a created submission is found to be invalid during processing.
+    
+    """
+    
+    if submission.testables:
+        # Delete the directory
+        shutil.rmtree(submission.testables)
 
 @app.route("/assignments/<assignment_id>/upload", methods = ["POST"])
 @account_type_required("student")
@@ -21,17 +64,42 @@ def upload_submission(assignment_id):
     try:
         id = ObjectId(assignment_id)
     except InvalidId:
-        app.logger.debug("Invalid ID")
+        app.logger.debug("Invalid ID: Malformed.")
         
         abort(404)
     
     # Ensure that an assignment with the provided id actually exists
     if Assignment.objects(id = id).limit(1).count() == 0:
+        app.logger.debug("Invalid ID: Assignment does not exist.")
+        
         abort(404)
     
+    def craft_response(**kwargs):
+        app.logger.debug(str(kwargs))
+        
+        if request.is_xhr:
+            # If the request was made via ajax return a JSON object...
+            return jsonify(**kwargs)
+        else:
+            # otherwise redirect to the correct view.
+            if "error" in kwargs:
+                flash(kwargs["error"], category = "error")
+            else:
+                flash("File(s) uploaded succesfully.", category = "message")
+                
+            redirect_to = request.args.get("next") or request.referrer
+            
+            if not is_url_on_site(app, redirect_to):
+                # Default going back to the assignment screen
+                redirect_to = url_for(
+                    "view_assignment", 
+                    assignment_id = assignment_id
+                )
+                
+            return redirect(redirect_to)
+    
     # Craft a new submission
-    new_submission = Submission(
-        id = ObjectId(),
+    new_submission = prepare_new_submission(
         assignment = id,
         user = current_user.id,
         timestamp = datetime.datetime.now()
@@ -41,54 +109,55 @@ def upload_submission(assignment_id):
     if request.files.get("archive"):
         archive = request.files["archive"]
         
-        # Store the archive somewhere on the filesystem so we can access it
+        # Create a temporary file that we will use to store the archive. It is
+        # given to us open and as a tuple with some additional information so we
+        # need to close it and extract only the information we need.
         temp_file = tempfile.mkstemp()
-        
-        # We don't want an open file handle to the temporary file so close it
         os.close(temp_file[0])
-        
-        # All we're interseted in is that path to the file
         temp_file = temp_file[1]
         
         # Save the archive over the temp_file we just created
         archive.save(temp_file)
         
-        # Figure out where we will store this submission
-        directory = os.path.join(SUBMISSION_DIRECTORY, str(new_submission.id))
-        
-        new_submission.testables = directory
-        
-        # Create the directories needed (this will try to create the entire
-        # directory tree if necessary).
-        os.makedirs(directory)
-        
         # TODO: Find out if tar is secure! Can the archive be made such that
         # files will be placed outside of the directory were exporting the
         # tar into?? If so that's a huge security hole.
-        # TODO: It would be nice if zip was supported. The reason it's not
-        # for now is that the unzip program won't accept an archie from
-        # standard input.
         try:
             if archive.filename.endswith(".tar"):
                 subprocess.check_call(
                     ["tar", "xf", temp_file], 
-                    cwd = directory
+                    cwd = new_submission.testables
                 )
             elif archive.filename.endswith(".tar.gz"):
                 subprocess.check_call(
                     ["tar", "xzf", temp_file], 
-                    cwd = directory
+                    cwd = new_submission.testables
+                )
+            elif archive.filename.endswith(".zip"):
+                subprocess.check_call(
+                    ["unzip", temp_file],
+                    cwd = new_submission.testables
                 )
             else:
-                return "Invalid Filetype"
+                abort_new_submission(new_submission)
+                
+                return craft_response(
+                    error = "File had an unrecognized extension."
+                )
         except subprocess.CalledProcessError:
-            # Delete the directory
-            shutil.rmtree(directory)
+            abort_new_submission(new_submission)
             
-            return "Invalid Filetype"
+            return craft_response(
+                error = "File could not be opened as an archive."
+            )
         finally:
-            os.remove(temp_file)
+            # Always remove the temporary file we used to store the archive if
+            # we succesfully created one
+            if temp_file:
+                os.remove(temp_file)
     else:
-        return "NTD"
+        # We did not recieve enough information to do anything
+        return craft_response(error = "Nothing to do.")
     
-    return "GOOD"
+    # Everything seems to have gone well
+    return craft_response()
